Enemy AI
This script is a basic "zombie script" which means that all this enemy does is follow the player and when reaching a certain distance, it will play an attack animation. For this script to work you will need to attach a NavMeshAgent to your enemy or NPC/Nonplayable character and have a NavMesh on your terrain.