ItemCollection and score display
This script adds the functionality of collecting items in order to win the game. When the player is in range of an item, the number of items collected will be increased, and it will delete the object that has been collected in the game wold, simulating the character actually collecting the item.